Moved class to another package #95
