+++ Server/src/main/java/net/hawkengine/services/UserGroupService.java
@@ -1,6 +1,6 @@
 package net.hawkengine.services;
 
-import net.hawkengine.core.utilities.EndpointConnector;
+import net.hawkengine.ws.EndpointConnector;
 import net.hawkengine.db.DbRepositoryFactory;
 import net.hawkengine.db.IDbRepository;
 import net.hawkengine.model.ServiceResult;
